Refactor Helper into UserHelper and EmailHelper
