+++ config/app.php
@@ -227,7 +227,8 @@ return [
         'URL' => Illuminate\Support\Facades\URL::class,
         'Validator' => Illuminate\Support\Facades\Validator::class,
         'View' => Illuminate\Support\Facades\View::class,
-        'Helper' => App\Helpers\Helper::class,
+        'UserHelper' => App\Helpers\UserHelper::class,
+        'EmailHelper' => App\Helpers\EmailHelper::class,
         // 'Debugbar' => Barryvdh\Debugbar\Facade::class,
     ],
