+++ resources/views/admin/deleted-users.blade.php
@@ -150,7 +150,7 @@
 
                                                                     <div class="col-md-6">
                                                                         <div class="mb-2">
-                                                                            <img src="{{ Helper::getProfileImage($user->profile_picture) }}" alt="Profile Picture" width="100" height="100">
+                                                                            <img src="{{ UserHelper::getProfileImage($user->profile_picture) }}" alt="Profile Picture" width="100" height="100">
                                                                         </div>
                                                                         {{-- <div class="mb-2">
                                                                             <strong>Type :</strong> {{ $user->type }}
