+++ resources/views/admin/profile-detail.blade.php
@@ -201,7 +201,7 @@
 
                 <div class="col-md-6">
                     <div class="mb-2">
-                        <img src="{{ Helper::getProfileImage($profilePost['userdetails']['profile_picture']) }}" alt="Profile Picture" width="100" height="100">
+                        <img src="{{ UserHelper::getProfileImage($profilePost['userdetails']['profile_picture']) }}" alt="Profile Picture" width="100" height="100">
                     </div>
                     {{-- <div class="mb-2">
                         <strong>Type :</strong> {{ $user->type }}
