+++ resources/views/partials/job_notification_provider.blade.php
@@ -8,7 +8,7 @@
                     <div class="row">
                         <div class="col-md-9 p-0">
                             <div class="provider-picy" >
-                                <img src="{{ Helper::getProfilePic($allUserJob['job_userid']) }}" style="border-radius: 14px;" alt="">
+                                <img src="{{ UserHelper::getProfilePic($allUserJob['job_userid']) }}" style="border-radius: 14px;" alt="">
                             </div>
                             <h3 class="mb-0"
                                 style="display: block;margin-left: 90px;padding-bottom: 27px;padding-top:27px;">
@@ -26,15 +26,15 @@
                                  @endif
                                  @if ($allUserJob['fullconnect'])
                                    
-                                     @if (Helper::getTotalCount($allUserJob['conid'],$allUserJob['job_userid']) == 0 && $allUserJob['requested_by'] == 'service_provider')
-                                         @if (Helper::getUnseenCount($allUserJob['conid'], $allUserJob['provider_id']) > 0)
+                                     @if (UserHelper::getTotalCount($allUserJob['conid'],$allUserJob['job_userid']) == 0 && $allUserJob['requested_by'] == 'service_provider')
+                                         @if (UserHelper::getUnseenCount($allUserJob['conid'], $allUserJob['provider_id']) > 0)
                                              {{ $allUserJob['name'] }}  has accepted your job application.<br> You have a New Message from {{ $allUserJob['name'] }}
                                          @else
                                              {{ $allUserJob['name'] }}  has accepted your job application.<br> Chat with {{ $allUserJob['name'] }}  now.
                                          @endif
-                                     @elseif (Helper::getTotalCount($allUserJob['conid'],$allUserJob['job_userid']) == 0 && $allUserJob['requested_by'] != 'service_provider')
+                                     @elseif (UserHelper::getTotalCount($allUserJob['conid'],$allUserJob['job_userid']) == 0 && $allUserJob['requested_by'] != 'service_provider')
                                          Chat with {{ $allUserJob['name'] }} now
-                                     @elseif (Helper::getUnseenCount($allUserJob['conid'], $allUserJob['provider_id']) > 0)
+                                     @elseif (UserHelper::getUnseenCount($allUserJob['conid'], $allUserJob['provider_id']) > 0)
                                          You have a New Message from {{ $allUserJob['name'] }}
                                      @endif
                                  @endif
@@ -60,7 +60,7 @@
                                                 >
 
                                                     <i class="fas fa-comment" style="background-color: transparent; width: 50px;">
-                                                        <span id="chatCount-{{ $allUserJob['conid'] }}">{{ Helper::getUnseenCount($allUserJob['conid'],$allUserJob['provider_id']) }}</span>
+                                                        <span id="chatCount-{{ $allUserJob['conid'] }}">{{ UserHelper::getUnseenCount($allUserJob['conid'],$allUserJob['provider_id']) }}</span>
                                                     </i>
                                                 </a>
                                             </li>
