+++ resources/views/user/provider/service-profiles/housekeeper.blade.php
@@ -492,7 +492,7 @@
     <script>
         var profileId = "{{ session('profileid') }}";
         var saved = "{{ session('saved') }}";
-        var countWorkProfiles = @json(Helper::countWorkProfiles());
+        var countWorkProfiles = @json(UserHelper::countWorkProfiles());
 
         fetch('/user/check-new')
             .then(response => response.json())
